Reorganize repo internals to prepare for rfc#230 implementation.
Moves all of the future deprecated code into a `legacy-0-6-x` subfolder
(while preserving import paths) in order to ease implementation of
emberjs/rfcs#230.
