+++ addon-test-support/index.js
@@ -1,11 +1,19 @@
-export { default as TestModule } from './test-module';
+export { default as TestModule } from './legacy-0-6-x/test-module';
 export {
   default as TestModuleForAcceptance,
-} from './test-module-for-acceptance';
-export { default as TestModuleForComponent } from './test-module-for-component';
-export { default as TestModuleForModel } from './test-module-for-model';
-export { getContext, setContext, unsetContext } from './test-context';
-export { setResolver } from './test-resolver';
+} from './legacy-0-6-x/test-module-for-acceptance';
+export {
+  default as TestModuleForComponent,
+} from './legacy-0-6-x/test-module-for-component';
+export {
+  default as TestModuleForModel,
+} from './legacy-0-6-x/test-module-for-model';
+export {
+  getContext,
+  setContext,
+  unsetContext,
+} from './legacy-0-6-x/test-context';
+export { setResolver } from './resolver';
 
 import Ember from 'ember';
 Ember.testing = true;
