+++ addon-test-support/legacy-0-6-x/abstract-test-module.js
@@ -1,8 +1,8 @@
 import { run } from '@ember/runloop';
 import { Promise as EmberPromise, resolve } from 'rsvp';
 import { assign, merge as emberMerge } from '@ember/polyfills';
-import { _setupPromiseListeners, _teardownPromiseListeners } from './ext/rsvp';
-import { _setupAJAXHooks, _teardownAJAXHooks } from './wait';
+import { _setupPromiseListeners, _teardownPromiseListeners } from '../ext/rsvp';
+import { _setupAJAXHooks, _teardownAJAXHooks } from '../wait';
 import { getContext, setContext, unsetContext } from './test-context';
 
 import Ember from 'ember';
