+++ addon-test-support/legacy-0-6-x/test-module-for-acceptance.js
