+++ addon-test-support/resolver.js
